Update present_on_mulitple_rows_within to use proper value counts within groupings
