+++ business_rules/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.3.7"
+__version__ = "1.4.0"
 
 from .engine import run_all
 from .utils import export_rule_data
